Support transitioning from a previous etcd release
This change adds the correct extended key usages for TLS certificates so
that mutual auth TLS can be properly performed.

This change moves the `single-node.yml` to a generic `etcd.yml`. We can
use operations files to change configuration on this generic manifest.
It also renames the deployment name to `etcd` so that we can easily
migrate from existing deployments.

This change uses a persistent_disk_type rather than persistent_disk.
This allows defining persistent disks in the cloud-config rather than
relying on the CPI's defaults.

[#155132619]