Add etcdctl helper script
This change adds an `/var/vcap/jobs/etcd/bin/etcdctl` helper script that
comes tempalted with client certificates and the correct endpoints. If
you specify `LOCAL=true` then the `etcdctl` will only talk to the
current node. By default, when creating a login shell, the helper script
will be places on the operators PATH.

This change also renames `ca.pem` to `server-ca.crt` due to the presence
of the `client-ca.crt`.

[#155132619]

Signed-off-by: James Myers <jfmyers9@gmail.com>
Signed-off-by: Ben Chapman <bchapman@pivotal.io>