Update bbr-etcd job to use new bbr backup_one_restore_all feature
bbr named artifacts feature, using metadata script, will soon be deprecated.
This commit updates the job to use the new backup_one_restore_all
feature which will replace the named artifacts. Since this is a new
feature the job should fail if the bbr version being run does not
support it.

We used spec tests to drive out the behaviour in the templates.

[#164144061]

Signed-off-by: Mehul Modha <mmodha@pivotal.io>