Make VM deletion check more explicit
This example is referenced in the official BOSH documentation for drain (see https://bosh.io/docs/drain/#environment-variables) so people are coming here for reference on how to add a VM deletion check to their drain scripts.

This also adds a `-q` to grep to silence the positive path of the if condition. This might break drain scripts easily if different ways of stdout and stderr redirection is used.